backend/requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Caching
redis==5.0.1
aioredis==2.0.1

# Core API
openai>=1.10.0,<2.0.0
langchain==0.1.0
langchain-openai==0.0.5

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.12.0
ruff==0.1.8
mypy==1.7.1

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
numpy==1.26.2
pandas==2.1.3

# Logging & Monitoring
python-json-logger==2.0.7

# CORS & Security
fastapi-cors==0.0.6
